requirements.txt
numpy
opencv-python
pandas
Pillow
scipy
opencv-python
matplotlib